# openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageviewRequest": {
        "properties": {
          "api_key": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "referrer": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referrer"
          },
          "session_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "api_key",
          "session_id",
          "url"
        ],
        "title": "PageviewRequest",
        "type": "object"
      },
...
      "TrackEventRequest": {
        "properties": {
          "api_key": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "event_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Event Name",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "session_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "api_key",
          "event_name",
          "session_id"
        ],
        "title": "TrackEventRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Internal-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Self-hosted, privacy-first analytics platform controlled entirely through a Telegram bot.",
    "title": "tgram-analytics",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pageview": {
      "post": {
        "description": "Ingest a pageview event.\n\nForces ``event_name = \"pageview\"`` and stores url/referrer in dedicated\ncolumns as well as in ``properties`` for easy querying.",
        "operationId": "pageview_api_v1_pageview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pageview Api V1 Pageview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pageview",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/track": {
      "post": {
        "description": "Ingest a custom event.\n\nReturns 202 immediately; alert evaluation runs as a background task.",
        "operationId": "track_api_v1_track_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Track Api V1 Track Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  }
}